restructure code a tad bit
