+++ src/plugins/dashboard/public/dashboard_app/top_nav/add_new_panel/add_panel_action_menu_items.test.ts
