+++ src/plugins/dashboard/public/dashboard_app/top_nav/add_new_panel/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export { useGetDashboardPanels } from './use_get_dashboard_panels';
+export { DashboardPanelSelectionListFlyout } from './dashboard_panel_selection_flyout';
